Split into multiple files for better handling. Create Changelog.
