+++ changelogs/fragments/site_management.yml
@@ -1,0 +1,3 @@
+major_changes:
+  - Site management module - Added a module for site_management.
+    Refer to the module documentation for further details.
